\name{add.ExtObject}
\alias{add.ExtObject}
\title{method for add...}
\usage{\method{add}{ExtObject}(x, child, ...)
}
\description{method for add}
\arguments{\item{x}{object}
\item{child}{child object}
\item{...}{Not implemented. For values like \code{expand}, \code{fill} and \code{anchor}
S3 method for add and ExtObject}
}

\alias{add.ExtObject}

