% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add.GComponent}
\alias{add.GComponent}
\title{method for add}
\usage{
\method{add}{GComponent}(x, child, ...)
}
\arguments{
\item{x}{object}

\item{child}{child object}

\item{...}{Not implemented. For values like \code{expand}, \code{fill} and \code{anchor}
S3 method for add and GComponent}
}
\description{
method for add
}

