\name{addHandlerColumnClicked}
\alias{addHandlerColumnClicked}
\title{Assign handler to click event for column (gdf, gtable)...}
\usage{addHandlerColumnClicked(x, ...)
}
\description{Assign handler to click event for column (gdf, gtable)}
\value{a callback id, used with \code{removeHandler}\code{addHandlerColumnClicked}: a callback id, used with \code{removeHandler}


}
\arguments{\item{x}{the object}
\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}

\alias{addHandlerColumnClicked}

