% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addHandlerDestroy}
\alias{addHandlerDestroy}
\title{Assign handler to destroy event (such as page being unloaded)}
\usage{
addHandlerDestroy(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}
\value{
a callback id, used with \code{removeHandler}
}
\description{
Assign handler to destroy event (such as page being unloaded)
}

