% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{addHandlerFocus}
\alias{addHandlerFocus}
\title{Assign handler to focus in event}
\usage{
addHandlerFocus(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}
\value{
a callback id, used with \code{removeHandler}
}
\description{
Assign handler to focus in event
}

