\name{addHandlerKeystroke}
\alias{addHandlerKeystroke}
\title{Assign handler to keystroke event...}
\usage{addHandlerKeystroke(x, ...)
}
\description{Assign handler to keystroke event}
\details{This handler gets passed back the key that was pressed in the
\code{key} component.\code{addHandlerKeystroke}: This handler gets passed back the key that was pressed in the
\code{key} component.


}
\value{a callback id, used with \code{removeHandler}\code{addHandlerKeystroke}: a callback id, used with \code{removeHandler}


}
\arguments{\item{x}{the object}
\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}

\alias{addHandlerKeystroke}

