\name{addHandlerMouseMotion}
\alias{addHandlerMouseMotion}
\title{Assign handler to mouse motion event...}
\usage{addHandlerMouseMotion(x, ...)
}
\description{Assign handler to mouse motion event}
\details{This handler gets passed back when the mouse moves. This can be really slow!\code{addHandlerMouseMotion}: This handler gets passed back when the mouse moves. This can be really slow!


}
\value{a callback id, used with \code{removeHandler}\code{addHandlerMouseMotion}: a callback id, used with \code{removeHandler}


}
\arguments{\item{x}{the object}
\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}

\alias{addHandlerMouseMotion}

