% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addHandlerSelectionChanged}
\alias{addHandlerSelectionChanged}
\title{Assign handler to a change of selection event}
\usage{
addHandlerSelectionChanged(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}
\description{
Some widgets have selection, this allows a callback to be changed when that occurs
}

