% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{addSpring.default}
\alias{addSpring.default}
\title{Method for addSpring}
\usage{
\method{addSpring}{default}(x)
}
\arguments{
\item{x}{a box container (e.g., \code{ggroup})}
}
\description{
Method for addSpring
}

