% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{addSpring.GGroup}
\alias{addSpring.GGroup}
\title{Method for addSpring}
\usage{
\method{addSpring}{GGroup}(x)
}
\arguments{
\item{x}{a box container (e.g., \code{ggroup})}
}
\description{
Method for addSpring
}

