% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{addSpring}
\alias{addSpring}
\title{Add a spring to push additionnal child objects to far end of packing}
\usage{
addSpring(x)
}
\arguments{
\item{x}{a box container (e.g., \code{ggroup})}
}
\description{
For box Containers only
}

