% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/array.R
\docType{class}
\name{Array-class}
\alias{Array}
\alias{Array-class}
\title{An array class.}
\format{An \code{\link{R6Class}} generator object}
\usage{
Array
}
\description{
Like a list, but has some methods. Completely superflous, but
makes copying some code algorithms easier. We implement methods
such as \code{append}, \code{push}, \code{pop} and \code{each} for
iteration. As well, there are some lookup methods.
}
\section{Fields}{

\describe{
\item{\code{l}}{The list where the key value pairs are stored}

\item{\code{id_ctr}}{An integer counter for getting new names (as character strings)}
}}
\examples{
A <- Array$new(a = 2, b = 2)
}
\keyword{datasets}

