\name{asURL}
\alias{asURL}
\title{Add URL to class of object if not already...}
\usage{asURL(x)
}
\description{Add URL to class of object if not already}
\value{returns object\code{asURL}: returns object


}
\arguments{\item{x}{object to add class to. Should be length 1 character}
}

\alias{asURL}

