\name{blockHandler}
\alias{blockHandler}
\title{temporariliy block handler by id}
\usage{
  blockHandler(x, cbid)
}
\arguments{
  \item{x}{the object handler is called on.}

  \item{cbid}{the callback id returned by an addHandlerXXX
  call}
}
\value{
  NULL
}
\description{
  temporariliy block handler by id
}
\note{
  see the reference method \code{block_handlers} to block
  all handlers for the widget.
}

