% Generated by roxygen2 (4.0.1): do not edit by hand
\name{[.GComponent}
\alias{[.GComponent}
\title{Method for [}
\usage{
\method{[}{GComponent}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{object}

\item{i}{row index}

\item{j}{column index}

\item{...}{passed to \code{get_items}}

\item{drop}{passed to \code{get_items}}
}
\description{
Method for [
}

