\name{[<-.GComponent}
\alias{[<-.GComponent}
\title{assignment method for "["}
\usage{
  \method{"["}{GComponent} (x, i, j, ..., value) <- value
}
\arguments{
  \item{x}{objecct}

  \item{i}{row}

  \item{j}{column}

  \item{...}{passed to \code{set_items}}

  \item{value}{passed to \code{set_items}}
}
\description{
  assignment method for "["
}

