% Generated by roxygen2 (4.0.2): do not edit by hand
\name{browse_df}
\alias{browse_df}
\title{Browse a data frame in the global environment using gtable}
\usage{
browse_df(d, width = 800, height = 500, paging = nrow(d) > 200, ...)
}
\arguments{
\item{d}{data frame}

\item{width}{width of display}

\item{height}{height of display}

\item{paging}{Logical. Do we have paging to display values. Should use if \code{nrow(d) > 200} or so}

\item{...}{ignored}
}
\value{
creates webpage showing data frame
}
\description{
Browse a data frame in the global environment using gtable
}
\examples{
\dontrun{
browse_df(mtcars)
}
}

