% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coerceToJSString}
\alias{coerceToJSString}
\alias{coerceToJSString.String}
\alias{coerceToJSString.character}
\alias{coerceToJSString.default}
\alias{coerceToJSString.factor}
\alias{coerceToJSString.list}
\alias{coerceToJSString.logical}
\title{Coerce an object to a string}
\usage{
coerceToJSString(x)

\method{coerceToJSString}{default}(x)

\method{coerceToJSString}{character}(x)

\method{coerceToJSString}{factor}(x)

\method{coerceToJSString}{logical}(x)

\method{coerceToJSString}{String}(x)

\method{coerceToJSString}{list}(x)
}
\arguments{
\item{x}{object to coerce}
}
\value{
returns object sometimes quoted
}
\description{
Coerce an object to a string

S3 method for coerceToString

S3 method for coerceToString

S3 method for coerceToString

S3 method for coerceToString

S3 method for coerceToString

S3 method for coerceToString
}

