\name{coerceValue.numeric}
\alias{coerceValue.numeric}
\title{method to coerce value before assigning into data frame...}
\usage{coerceValue.numeric(x, value)
}
\description{method to coerce value before assigning into data frame}
\arguments{\item{x}{what value will go into}
\item{value}{to set into x}
}

