\name{coerceValue.integer}
\alias{coerceValue.integer}
\title{method to coerce value before assigning into data frame}
\usage{
  coerceValue.integer(x, value)
}
\arguments{
  \item{x}{what value will go into}

  \item{value}{to set into x}
}
\description{
  method to coerce value before assigning into data frame
}

