\name{column_editor.default}
\alias{column_editor.default}
\title{Method to create column editor for a table widget...}
\usage{column_editor.default(x, ...)
}
\description{Method to create column editor for a table widget}
\arguments{\item{x}{type of object}
\item{...}{ignored}
}

\alias{column_editor.default}

