% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\name{column_editor.default}
\alias{column_editor.default}
\title{Method to create column editor for a table widget}
\usage{
\method{column_editor}{default}(x, ...)
}
\arguments{
\item{x}{type of object}

\item{...}{ignored}
}
\description{
Method to create column editor for a table widget
}

