% Generated by roxygen2 (4.0.2): do not edit by hand
\name{column_editor}
\alias{column_editor}
\title{Generic function to create column editor for a table widget}
\usage{
column_editor(x, ...)
}
\arguments{
\item{x}{type of object}

\item{...}{ignored}
}
\description{
Generic function to create column editor for a table widget
}

