% Generated by roxygen2 (4.0.2): do not edit by hand
\name{column_renderer.default}
\alias{column_renderer.default}
\title{Method to render a column in ExtJS Grid}
\usage{
\method{column_renderer}{default}(x)
}
\arguments{
\item{x}{object to get column class from}
}
\description{
Method to render a column in ExtJS Grid
}

