% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\name{column_renderer.Icon}
\alias{column_renderer.Icon}
\title{Method to render a column in ExtJS Grid}
\usage{
\method{column_renderer}{Icon}(x)
}
\arguments{
\item{x}{object to get column class from}
}
\description{
Method to render a column in ExtJS Grid
}

