% Generated by roxygen2 (4.0.1): do not edit by hand
\name{column_renderer}
\alias{column_renderer}
\title{Generic function to render a column in ExtJS Grid}
\usage{
column_renderer(x)
}
\arguments{
\item{x}{object to get column class from}
}
\description{
Generic function to render a column in ExtJS Grid
}

