% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{delete.GComponent}
\alias{delete.GComponent}
\title{delete method}
\usage{
\method{delete}{GComponent}(x, child, ...)
}
\arguments{
\item{x}{object}

\item{child}{object}

\item{...}{ignored}
}
\description{
delete method
}

