% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{delete}
\alias{delete}
\title{Delete child object}
\usage{
delete(x, child, ...)
}
\arguments{
\item{x}{object}

\item{child}{child objecct}

\item{...}{ignored}
}
\description{
Delete child object
}

