% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{dispose.GComponent}
\alias{dispose.GComponent}
\title{Method for dispose}
\usage{
\method{dispose}{GComponent}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
Method for dispose
}

