% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{download_extjs}
\alias{download_extjs}
\title{download extjs files and unzips into location to be found}
\usage{
download_extjs()
}
\value{
Nothing. Installs files if successful
}
\description{
Useful if installing from a package which doesn't bundle. The
extjs files are quite large (nearly 70Mb download), so rather than
bundle in an R package, one can download as needed.
}

