\name{editable<-}
\alias{editable<-}
\title{Toggle editability of the object, if supported}
\usage{
  editable(x, ...) <- value
}
\arguments{
  \item{x}{object. Calls objects \code{set_editable}
  method}

  \item{...}{passed to \code{set_editable} method}

  \item{value}{logical}
}
\description{
  Toggle editability of the object, if supported
}

