\name{emptyJSArray}
\alias{emptyJSArray}
\title{An empty array...}
\usage{emptyJSArray(doBrackets=TRUE)
}
\description{An empty array}
\details{Used by toJSArray as helper function\code{emptyJSArray}: Used by toJSArray as helper function


}
\arguments{\item{doBrackets}{logical}
}

\alias{emptyJSArray}

