% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{enabled}
\alias{enabled}
\title{Is widget enabled, sensitive to user input}
\usage{
enabled(x)
}
\arguments{
\item{x}{the widget}
}
\description{
Is widget enabled, sensitive to user input
}

