\name{enabled<-}
\alias{enabled<-}
\alias{enabled<-.GComponent}
\title{Set if widget is enabled}
\usage{
  enabled(x, value) <- value

  \method{enabled}{GComponent} (x, value) <- value
}
\arguments{
  \item{x}{widget}

  \item{value}{logical}
}
\description{
  Set if widget is enabled

  assignment method for enabled
}

