% Generated by roxygen2 (4.0.2): do not edit by hand
\name{escapeSingleQuote}
\alias{escapeSingleQuote}
\title{escape single quote}
\usage{
escapeSingleQuote(x)
}
\arguments{
\item{x}{string to escape}
}
\value{
string with "'" replaced by "\'"
}
\description{
Useful as we often single quote arguments to the Ext functions
}

