% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ext-misc.R
\docType{class}
\name{ExtArgs-class}
\alias{ExtArgs}
\alias{ExtArgs-class}
\title{A class to handle map between R and arguments for constructor}
\format{An \code{\link{R6Class}} generator object}
\usage{
ExtArgs
}
\description{
The ExtArgs are simply an Array instance wrapper with an extend method and
a conversion method
}
\section{Fields}{

\describe{
\item{\code{args}}{\code{\link{Array}} of arguments}
}}
\keyword{datasets}

