\name{focus<-}
\alias{focus<-}
\alias{focus<-.default}
\title{Set focus onto object.}
\usage{
  focus(obj) <- value

  \method{focus}{default} (obj, value) <- value
}
\arguments{
  \item{value}{logical. Set focus state.}
}
\description{
  For some widgets, this sets user focus (e.g. gedit gets
  focus for typing).

  Basic S3 method for focus
}

