% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gaction.R
\docType{data}
\name{GAction}
\alias{GAction}
\title{Class for gaction}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x27bbd78> 
 - attr(*, "name")= chr "GAction_generator"
}}
\usage{
GAction
}
\description{
Class for gaction
}
\keyword{datasets}

