% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdialogs.R
\docType{class}
\name{GDialog-class}
\alias{GDialog}
\alias{GDialog-class}
\title{Base class for dialogs}
\description{
Base class for dialogs
}
\section{Methods}{

\describe{
\item{\code{add_R_callback(signal, ...)}}{Add a callback into for the Ext signal. Return callback idas a list.}
}}

