\name{get_tempfile_url}
\alias{get_tempfile_url}
\title{Get url for a tempfile created by get_tempfile...}
\usage{get_tempfile_url(f)
}
\description{Get url for a tempfile created by \code{get_tempfile}}
\details{Requires the tempdir to be mapped to a specific url\code{get_tempfile_url}: Requires the tempdir to be mapped to a specific url


}
\arguments{\item{f}{a file name produced by \code{get_tempfile}}
}
\examples{f <- get_tempfile(ext=".svg")
get_tempfile_url(f)\code{get_tempfile_url}: f <- get_tempfile(ext=".svg")
get_tempfile_url(f)


}

\alias{get_tempfile_url}

