% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_tempfile}
\alias{get_tempfile}
\title{Get a temporary file matching the given extension, if given.}
\usage{
get_tempfile(ext = ".txt")
}
\arguments{
\item{ext}{file extension}
}
\description{
This tempfile is stored in a directory that can be served as a static file through
gWidgetsWWW2.
}
\examples{
f <- get_tempfile(ext=".svg")
get_tempfile_url(f)
}

