% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getFromDots}
\alias{getFromDots}
\title{get value from ... arguments}
\usage{
getFromDots(key, ..., default = NULL)
}
\arguments{
\item{key}{key to lookip}

\item{...}{passed from function}

\item{default}{default value if NULL}
}
\description{
get value from ... arguments
}

