\name{getStockIconByName}
\alias{getStockIconByName}
\title{Get the icon class (or url) by its name...}
\usage{getStockIconByName(icon_name, css=TRUE)
}
\description{Get the icon class (or url) by its name}
\value{Icon css class (or url) or NA if not there\code{getStockIconByName}: Icon css class (or url) or NA if not there


}
\arguments{\item{icon_name}{icon name to be looked up in stock icons}
\item{css}{logical. If TRUE, return css class name, else url}
}
\examples{getStockIconByName("up") ## "x-gw-icon-up"
getStockIconByName("Not there") ## NULL\code{getStockIconByName}: getStockIconByName("up") ## "x-gw-icon-up"
getStockIconByName("Not there") ## NULL


}

\alias{getStockIconByName}

