\name{getStockIcons}
\alias{getStockIcons}
\title{Return stock icons}
\value{
  a list of icons. Unlike other gWidgets implementations.
  Each component is alist with css name and url, as both
  are useful.
}
\description{
  These appear as a list with a name, CSS class
  (setIconClass), and url (setIcon)
}

