\name{ggooglemaps_infowindow}
\alias{ggooglemaps_infowindow}
\title{Add an infowindow -- a popup cartoon -- to the map...}
\usage{ggooglemaps_infowindow(position, content, map)
}
\description{Add an infowindow -- a popup cartoon -- to the map}
\details{An info window allows one to display HTML code in a popup window\code{ggooglemaps_infowindow}: An info window allows one to display HTML code in a popup window


}
\value{a GGoogleMapsInfoWindow object\code{ggooglemaps_infowindow}: a GGoogleMapsInfoWindow object


}
\arguments{\item{position}{Either a \code{ggooglemaps_marker} object or position as lat/lng pair}
\item{content}{HTML fragment to display in window}
\item{map}{a \code{ggooglemaps} object}
}

\alias{ggooglemaps_infowindow}

