\name{ggooglemaps_polygon}
\alias{ggooglemaps_polygon}
\title{Draw a polygon on the map...}
\usage{ggooglemaps_polygon(path, stroke=list(color = "#0000FF", opacity = 1, weight = 2),
    fill=list(color = "#FF0000", opacity = 0.25), map, ...)
}
\description{Draw a polygon on the map}
\details{Draw a polygon specified as  matrix of data frame of latitude/longitude values\code{ggooglemaps_polygon}: Draw a polygon specified as  matrix of data frame of latitude/longitude values


}
\value{a polygon instance, but has no methods\code{ggooglemaps_polygon}: a polygon instance, but has no methods


}
\arguments{\item{path}{path of polygon as matrix or data frame of points}
\item{stroke}{list of stroke parameters (color, opacity, weight)}
\item{fill}{list of fill parameters (color, opacity)}
\item{map}{ggooglemaps object}
\item{...}{ignored}
}

\alias{ggooglemaps_polygon}

