\name{ggooglemaps_marker}
\alias{ggooglemaps_marker}
\title{Add a marker to the map}
\usage{
  ggooglemaps_marker(position, title = NULL, icon = NULL,
  map)
}
\arguments{
  \item{position}{lat/lng pair for position of mark}

  \item{title}{optional tooltip text}

  \item{icon}{option icon, stock name or url}

  \item{map}{ggooglemaps instance}
}
\value{
  a GGoogleMapsObject
}
\description{
  Add a marker to the map
}

