% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ggooglemaps.R
\docType{class}
\name{GGoogleMapsObject-class}
\alias{GGoogleMapsObject}
\alias{GGoogleMapsObject-class}
\title{base class for ggooglemaps objects}
\description{
base class for ggooglemaps objects
}
\section{Methods}{

\describe{
\item{\code{add_R_callback(signal, ...)}}{Add a callback into for the Ext signal. Return callback idas a list.}

\item{\code{write_constructor()}}{Write out constructor.}
}}

