% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ggooglemaps.R
\docType{class}
\name{GGoogleMapsPolygon-class}
\alias{GGoogleMapsPolygon}
\alias{GGoogleMapsPolygon-class}
\title{base class for ggooglemaps polygon}
\description{
base class for ggooglemaps polygon
}

