\name{ghtml}
\alias{ghtml}
\title{widget to render HTML text pages...}
\usage{ghtml(x, container, ..., width, height, ext.args)
}
\description{widget to render HTML text pages}
\value{an \code{ExtWidget} object\code{ghtml}: an \code{ExtWidget} object


}
\arguments{\item{x}{an HTML string or a URL.}
\item{container}{parent container}
\item{...}{passed to add method of parent container}
\item{width}{width of widge in pixels}
\item{height}{height of widget in pixels}
\item{ext.args}{extra arguments to pass to constructor}
}
\examples{w <- gwindow()
h <- ghtml("<b>this is bold</b>", cont=w)\code{ghtml}: w <- gwindow()
h <- ghtml("<b>this is bold</b>", cont=w)


}

\alias{ghtml}

