% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdialogs.R
\name{ginput}
\alias{ginput}
\title{input dialog.}
\usage{
ginput(message, text = "", title = "Input", icon = NULL, parent = NULL,
  handler = NULL, action = NULL, ...)
}
\arguments{
\item{message}{message}

\item{text}{initial text for the widget}

\item{title}{title for dialog's window}

\item{icon}{icon ignored here}

\item{parent}{parent container (main window instance)}

\item{handler}{Called if yes is selected, the component
\code{input} of the first argument holds the user-supplied string.}

\item{action}{passed to any handler}

\item{...}{ignored}
}
\value{
return value ignored, use handler for response
}
\description{
Used for getting a text string to pass to a handler. Unlike other
gWidgets implementations, this call does not block the R process,
so any response to the user must be done through the handler.
}
\examples{
w <- gwindow()
gbutton("click me for a message", cont=w, handler=function(h,...) {
  ginput("What is your name?", parent=w, handler=function(h,...) {
    galert(paste("Hello", h$input), parent=w)
  })
})
}

