% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{gmenu-class}
\alias{GMenu}
\alias{gmenu-class}
\title{base class for menu instances.}
\description{
base class for menu instances.
}
\section{Methods}{

\describe{
\item{\code{remove_from_toolbar(obj)}}{Remove object from toolbar}
}}

