\name{GSlider}
\alias{GSlider}
\title{base class for gslider}
\description{
  The slider implementation allows one to adjust the
  formating of the tooltip that indicates the current value
  of the widget. The \code{tpl} command uses an HTML
  snippet where the value \code{{0}} does the substitution.
}

