% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gstatusbar.R
\docType{class}
\name{gstatusbar-class}
\alias{GStatusbar}
\alias{gstatusbar-class}
\title{\code{GStatusbar}  class for gstatusbar}
\description{
The convenience reference class method \code{clear} will clear the
text. This is basically the same as \code{svalue(obj) <- ""}.
}
\details{
We add in reference methods \code{show_loading} and
\code{hide_loading} to allow one to indicate something is
loading. Not really that useful though, as is. This is due to how
the JavaScript commands come back from the R process in a block so
unless the browser takes a long time to process the returning
JavaScript, you will likely show the loading for just a bit.

This class inherits for \code{GContainer}, thereby allowing one to
add in widgets in addition to the message by using the status bar
as the parent container. This might be useful.
}
\section{Methods}{

\describe{
\item{\code{get_value(...)}}{Get main property, Can't query widget, so we store here}

\item{\code{set_value(value, ...)}}{Set main property, invoke change handler on change}
}}

