% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{class}
\name{GWidgetArrayProxy-class}
\alias{GWidgetArrayProxy}
\alias{GWidgetArrayProxy-class}
\title{Proxy to return array data (combobox, table, gdf, ...) [[],[],[]]}
\description{
Proxy to return array data (combobox, table, gdf, ...) [[],[],[]]
}
\section{Methods}{

\describe{
\item{\code{get_data(...)}}{Return raw data, not JSON encoded}

\item{\code{get_fields()}}{Return json with fields mapped to names}

\item{\code{get_json_data(...)}}{The params determine what to pass back. }

\item{\code{post_json_data(param)}}{A post request from updating a store}

\item{\code{remove_row(param)}}{Remove the row}

\item{\code{set_data(value, ...)}}{Set proxy data}
}}

