% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{data}
\name{GWidgetArrayProxy}
\alias{GWidgetArrayProxy}
\title{Proxy to return array data (combobox, table, gdf, ...) [[],[],[]]}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x2e829c8> 
 - attr(*, "name")= chr "GWidgetArrayProxy_generator"
}}
\usage{
GWidgetArrayProxy
}
\description{
Proxy to return array data (combobox, table, gdf, ...) [[],[],[]]
}
\keyword{datasets}

