% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{class}
\name{GWidgetArrayStore-class}
\alias{GWidgetArrayStore}
\alias{GWidgetArrayStore-class}
\title{A store for an array (inbetween widget and proxy)}
\description{
A store for an array (inbetween widget and proxy)
}
\section{Methods}{

\describe{
\item{\code{init(value, ...)}}{Initialize proxy, then set up store}

\item{\code{write_model()}}{Write out model}
}}

