% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{data}
\name{GWidgetArrayStore}
\alias{GWidgetArrayStore}
\title{A store for an array (inbetween widget and proxy)}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x3857000> 
 - attr(*, "name")= chr "GWidgetArrayStore_generator"
}}
\usage{
GWidgetArrayStore
}
\description{
A store for an array (inbetween widget and proxy)
}
\keyword{datasets}

