\name{GWidgetGrid}
\alias{GWidgetGrid}
\title{A class inherited by GTable and GDf
Comment about \code{visible<-} being implemented partially through \code{$filter}}
\description{
  A class inherited by GTable and GDf Comment about
  \code{visible<-} being implemented partially through
  \code{$filter}
}

