% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{data}
\name{GWidgetHTMLProxy}
\alias{GWidgetHTMLProxy}
\title{Proxy for ghtml}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x357a490> 
 - attr(*, "name")= chr "GWidgetHTMLProxy_generator"
}}
\usage{
GWidgetHTMLProxy
}
\description{
Proxy for ghtml
}
\keyword{datasets}

