% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{class}
\name{GWidgetProxy-class}
\alias{GWidgetProxy}
\alias{GWidgetProxy-class}
\title{interface between R objects and JSON to supply Ext objects with data}
\description{
interface between R objects and JSON to supply Ext objects with data
}
\section{Methods}{

\describe{
\item{\code{get_data(...)}}{Return raw data, not JSON encoded}

\item{\code{get_fields()}}{Return json with fields mapped to names}

\item{\code{get_json_data(...)}}{The params determine what to pass back. }

\item{\code{get_url_list()}}{Get url to call back into this proxy object. The values 'proxy_url' and 'session_id' are JS globals in webpage}

\item{\code{set_data(value, ...)}}{Set proxy data}

\item{\code{set_value(value, ...)}}{Set main property, invoke change handler on change}
}}

