% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{data}
\name{GWidgetProxy}
\alias{GWidgetProxy}
\title{interface between R objects and JSON to supply Ext objects with data}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x2f40890> 
 - attr(*, "name")= chr "GWidgetProxy_generator"
}}
\usage{
GWidgetProxy
}
\description{
interface between R objects and JSON to supply Ext objects with data
}
\keyword{datasets}

