\name{GWidget}
\alias{GWidget}
\title{GWidget is the Base class for widget objects. See
\code{gcontainer} for the base class for container objects. Both
derive from the \code{GComponent} class.}
\description{
  GWidget is the Base class for widget objects. See
  \code{gcontainer} for the base class for container
  objects. Both derive from the \code{GComponent} class.
}

