% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GWidgetStore-class}
\alias{GWidgetStore}
\alias{GWidgetStore-class}
\title{Basic store, no default proxy so this isn't useful without being subclassed}
\description{
Basic store, no default proxy so this isn't useful without being subclassed
}
\section{Methods}{

\describe{
\item{\code{get_data(...)}}{Get data in store}

\item{\code{set_data(value, ...)}}{Set data for store}

\item{\code{write_model()}}{Write out model}
}}

