% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\docType{data}
\name{GWidgetStore}
\alias{GWidgetStore}
\title{Basic store, no default proxy so this isn't useful without being subclassed}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x3808d08> 
 - attr(*, "name")= chr "GWidgetStore_generator"
}}
\usage{
GWidgetStore
}
\description{
Basic store, no default proxy so this isn't useful without being subclassed
}
\keyword{datasets}

