\name{gWidgetsWWW2-generics}
\alias{coerceToJSString}
\title{Coerce an object to a string...}
\usage{coerceToJSString(x)
}
\description{Coerce an object to a string}
\value{returns object sometimes quoted\code{coerceToJSString}: returns object sometimes quoted


}
\arguments{\item{x}{object to coerce}
}

\alias{coerceToJSString}

