\name{gWidgetsWWW2-package}
\alias{gWidgetsWWW2-package}
\alias{gWidgetsWWW2-package}
\title{An implementation of the gWidgets API for web programming...}
\description{An implementation of the gWidgets API for web programming}
\details{This package uses the \pkg{Rook} package of Jeffrey Horner to
allow R programmers to easily create interactive web pages from
within an R session.\code{gWidgetsWWW2-package}: This package uses the \pkg{Rook} package of Jeffrey Horner to
allow R programmers to easily create interactive web pages from
within an R session.


}
\docType{packagepackage}
\author{John Verzani \email{jverzani@gmail.com}}
\keyword{package}

\alias{gWidgetsWWW2-package}
\alias{gWidgetsWWW2-package}
\keyword{package}

