% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GWidgetText-class}
\alias{GWidgetText}
\alias{GWidgetText-class}
\title{Base class for text widgets}
\description{
Base class for text widgets
}
\section{Methods}{

\describe{
\item{\code{param_defn(signal)}}{Define different parameter definitions based on the signal}

\item{\code{set_value(value, ...)}}{Set main property, invoke change handler on change}

\item{\code{transport_fun()}}{javascript function for transport web -> R. Creates an object param.
This is a string to be passed to the javascript queue withing the transport function call
E.g. var param = {value: this.getText()}}
}}

