\name{insert.ExtObject}
\alias{insert.ExtObject}
\title{method for insert...}
\usage{\method{insert}{ExtObject}(obj, value, where="end", font.attr, do.newline=TRUE, ...)
}
\description{method for insert}
\arguments{\item{obj}{object}
\item{value}{value to insert}
\item{where}{to insert}
\item{font.attr}{ignored}
\item{do.newline}{logical}
\item{...}{ignored}
}

