\name{insert.GComponent}
\alias{insert.GComponent}
\title{method for insert}
\usage{
  \method{insert}{GComponent} (obj, value, where = "end",
  font.attr = NULL, do.newline = TRUE, ...)
}
\arguments{
  \item{obj}{object}

  \item{value}{value to insert}

  \item{where}{to insert}

  \item{font.attr}{ignored}

  \item{do.newline}{logical}

  \item{...}{ignored}
}
\description{
  method for insert
}

