% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{is_running_local}
\alias{is_running_local}
\title{Is the process running locally?}
\usage{
is_running_local()
}
\value{
logical value, \code{TRUE} if local, \code{FALSE} if not.
}
\description{
Is the process running locally?
}

