% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isURL}
\alias{isURL}
\title{Is value a URL: either of our class URL or matches url string: ftp://, http:// or file:///}
\usage{
isURL(x)
}
\arguments{
\item{x}{length 1 character value to test}
}
\value{
Logical indicating if a URL.
}
\description{
Is value a URL: either of our class URL or matches url string: ftp://, http:// or file:///
}

