\name{JsonRPCServer}
\alias{JsonRPCServer}
\title{A JSON-RPC class
we expect a callback into the R session that looks like:}
\arguments{
  \item{obj}{name of aan JSON-rpc object in the global
  namespace}

  \item{method}{name an exported method of obj}

  \item{JSON}{encoded parameters to pass to the method call
  jsonRpc([{obj, method, params},{,,}]) we return JSON (or
  script)}
}
\description{
  A JSON-RPC class we expect a callback into the R session
  that looks like:
}

