% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\name{load_dirapp}
\alias{load_dirapp}
\title{Load an app in a directory}
\usage{
load_dirapp(dir_name, ...)
}
\arguments{
\item{dir_name}{The directory name (with full path). We make several assuptions:
1) \code{dir/*.R} has one answer, or we
take \code{index.R}, or we take first alphabetically; 2)
\code{app_name} is \code{basename(dir_name)}; 3)
\code{brew/*rhtml} is empty (fullscreen), has one value, or has
\code{index.rhtml} or we take first alphabetically; 4)
authenticator (XXX to be determined, for now calls
\code{make_session_manager})}

\item{...}{passed to \code{load_dir} reference class method of
\code{R_http} instance.}
}
\value{
creates the app
}
\description{
Load an app in a directory
}

