% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{make_rapache_files}
\alias{make_rapache_files}
\title{Make a template for the rapache configuration files}
\usage{
make_rapache_files(file = "")
}
\arguments{
\item{file}{passed to \code{cat} function to control what happens to output}
}
\description{
To use rapache a few locations need to be defined so that ajax
calls and calls to get local files are recognized. This happens in
a configuration file and this function makes a template that will
work, yet allows ucustomization.
}
\note{
The resulting file must be placed into the apached directory (/etc/apache2?)
and integrated into the configuration system.
}

