% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\name{makeFields}
\alias{makeFields}
\title{helper function to write field names for Ext constructor}
\usage{
makeFields(df)
}
\arguments{
\item{df}{data frame}
}
\description{
helper function to write field names for Ext constructor
}

