% Generated by roxygen2 (4.0.1): do not edit by hand
\name{merge.list}
\alias{merge.list}
\title{merge two lists}
\usage{
\method{merge}{list}(x, y, overwrite = TRUE)
}
\arguments{
\item{x}{a list}

\item{y}{a list}

\item{overwrite}{logical should we overright values in x}
}
\description{
merge two lists
}

