% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{names<-.GComponent}
\alias{names<-.GComponent}
\alias{namesGComponent}
\title{assignment method for names}
\usage{
\method{names}{GComponent}(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new names}
}
\description{
assignment method for names
}

