% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{ourQuote}
\alias{ourQuote}
\title{make a string safe to pass in as HTML fragment.}
\usage{
ourQuote(x)
}
\arguments{
\item{x}{a string to replace ' with}
}
\value{
a string with quotes escaped and placed within ''
}
\description{
We pass in strings as 'string contents' so we replace all ' quotes with " ones, then wrap in ''
}

