% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load.R
\docType{class}
\name{R_http}
\alias{R_http}
\title{The help server as a singleton R6 class}
\format{An \code{\link{R6Class}} generator object}
\usage{
R_http
}
\description{
The following code creates an instance of the R6 class R_http,
while not giving a name to its generator object, effectively
making the class a singleton within an instance of R.
}
\section{Fields}{

\describe{
\item{\code{R}}{An Rhttpd object}

\item{\code{port}}{The port of the Dynamic help server}

\item{\code{started}}{Is the R_http instance running?}

\item{\code{loaded}}{Is the gWidgets infrastructure loaded?}
}}
\keyword{datasets}

