\name{removeHandler}
\alias{removeHandler}
\title{Remove handler}
\usage{
  removeHandler(x, cbid)
}
\arguments{
  \item{x}{the object handler is called on.}

  \item{cbid}{the callback id returned by an addHandlerXXX
  call}
}
\value{
  NULL
}
\description{
  Remove handler
}
\note{
  see the reference method \code{remove_handlers} to remove
  all handlers on a widget.
}

