% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{Singleton}
\alias{Singleton}
\title{Singleton pattern}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7d58888> 
 - attr(*, "name")= chr "Singleton_generator"
}}
\usage{
Singleton
}
\description{
Singleton pattern
}
\keyword{datasets}

