\name{size_assign}
\alias{size<-}
\title{Set size property, if implemented...}
\usage{size(x, ...) <- value}
\description{Set size property, if implemented}
\arguments{\item{x}{widget}
\item{...}{passed on}
\item{value}{size specification, for most widgets a pair c(width, height), but can have exceptions}
}

\alias{size<-}

