\name{String}
\alias{String}
\title{String class -- does not get escaped in object literals...}
\usage{String(x)
}
\description{String class -- does not get escaped in object literals}
\details{Useful in the toJSObject method}
\value{vector with new class addedd}
\arguments{\item{x}{character vector to coerce to String class}
}

