\name{svalue_assign}
\alias{svalue<-}
\title{Set main value associated with a widget...}
\usage{svalue(x, ...) <- value}
\description{Set main value associated with a widget}
\arguments{\item{x}{object}
\item{...}{passed to \code{set_value} method. May include arguments for \code{index}}
\item{value}{value to set}
}

\alias{svalue<-}

