\name{svalue.GComponent}
\alias{svalue.GComponent}
\title{svalue method}
\usage{
  \method{svalue}{GComponent} (x, index = NULL,
    drop = NULL, ...)
}
\arguments{
  \item{x}{the widget}

  \item{...}{passed to \code{get_value} method. May include
  arguments \code{index} or \code{drop}}
}
\description{
  svalue method
}

