\name{svalue}
\alias{svalue}
\title{Return main value associated with a widget}
\usage{
  svalue(x, index = NULL, drop = NULL, ...)
}
\arguments{
  \item{x}{the widget}

  \item{...}{passed to \code{get_value} or \code{get_index}
  method. May include arguments \code{index} or
  \code{drop}}
}
\description{
  Return main value associated with a widget
}

