\name{svalue<-}
\alias{svalue<-}
\title{Set main value associated with a widget}
\usage{
  svalue(x, ...) <- value
}
\arguments{
  \item{x}{object}

  \item{...}{passed to \code{set_value} method. May include
  arguments for \code{index}}

  \item{value}{value to set}
}
\description{
  Set main value associated with a widget
}

