% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{tag.GComponent}
\alias{tag.GComponent}
\title{method for tag}
\usage{
\method{tag}{GComponent}(x, key)
}
\arguments{
\item{x}{the object}

\item{key}{key holding attribute}
}
\description{
method for tag
}

