% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{tag}
\alias{tag}
\title{Retreive a persistent attribute for an object}
\usage{
tag(x, key)
}
\arguments{
\item{x}{the object}

\item{key}{key holding attribute}
}
\description{
Retreive a persistent attribute for an object
}

