% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{tag<-}
\alias{tag<-}
\alias{tag<-.GComponent}
\title{Set a persistent attirbute for an object}
\usage{
tag(x, key) <- value

\method{tag}{GComponent}(x, key) <- value
}
\arguments{
\item{x}{object}

\item{key}{key to store the value}

\item{value}{attribute value}
}
\description{
Set a persistent attirbute for an object

assignment method for tag
}

