\name{toJSArray}
\alias{toJSArray}
\title{Make a JS array from an R object}
\usage{
  toJSArray(x, doBrackets = TRUE)
}
\arguments{
  \item{x}{R object to make into an array}

  \item{doBrackets}{logical Use brackets in ouput []}
}
\value{
  JSON encoded value from arrary
}
\description{
  Make a JS array from an R object
}

