% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tooltip<-}
\alias{tooltip<-}
\alias{tooltip<-.default}
\title{Set a tooltip for the widget}
\usage{
tooltip(x) <- value

\method{tooltip}{default}(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{character tooltip value}
}
\description{
Set a tooltip for the widget

Basic S3 method for tooltip<-
}

