\name{unblockHandler}
\alias{unblockHandler}
\title{unblock handler by id}
\usage{
  unblockHandler(x, cbid)
}
\arguments{
  \item{x}{the object handler is called on.}

  \item{cbid}{the callback id returned by an addHandlerXXX
  call}
}
\value{
  NULL
}
\description{
  unblock handler by id
}
\note{
  see the reference method \code{unblock_handlers} to
  unblock all the handlers for a widget.
}

