% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{update.GComponent}
\alias{update.GComponent}
\title{method for update}
\usage{
\method{update}{GComponent}(object, ...)
}
\arguments{
\item{object}{object}

\item{...}{passed along}
}
\description{
method for update
}

