\name{visible_assign}
\alias{visible<-}
\title{Primarily used to set if widget is shown, but also has other meanings...}
\usage{visible(x) <- value}
\description{Primarily used to set if widget is shown, but also has other meanings}
\arguments{\item{x}{object Calls objects \code{set_visible} method}
\item{value}{logical}
}

