\name{visible}
\alias{visible}
\alias{visible.GComponent}
\title{Generically returns if widget is in visible state}
\usage{
  visible(x)

  \method{visible}{GComponent} (x)
}
\arguments{
  \item{x}{object Calls objects \code{set_visible} method}
}
\description{
  Generically returns if widget is in visible state

  getters method for visible
}

