\name{add}
\alias{add}
\title{Add child component to parent container...}
\usage{add(x, child, ...)
}
\description{Add child component to parent container}
\details{Usually called internally by specifying \code{container=parent} to the constructor.\code{add}: Usually called internally by specifying \code{container=parent} to the constructor.


}
\value{nothing\code{add}: nothing


}
\arguments{\item{x}{object}
\item{child}{child object}
\item{...}{Not implemented. For values like \code{expand}, \code{fill} and \code{anchor}}
}

\alias{add}

