% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{addHandlerChanged}
\alias{addHandlerChanged}
\title{Assign handler to default event}
\usage{
addHandlerChanged(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}
\description{
The default event is specified in the definition of the
object. Using this function allows one to have a "generic" name
for adding a handler.
}

