% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{addHandlerKeystroke}
\alias{addHandlerKeystroke}
\title{Assign handler to keystroke event}
\usage{
addHandlerKeystroke(x, ...)
}
\arguments{
\item{x}{the object}

\item{...}{Used to pass through \code{handler}, and \code{action}
values. The handler is a function whose first argument is a list
that contains components \code{obj} to return the object and
\code{action}, holding the action value, and possibly others.}
}
\value{
a callback id, used with \code{removeHandler}
}
\description{
This handler gets passed back the key that was pressed in the
\code{key} component.
}

