\name{asIcon}
\alias{asIcon}
\title{Mark a character as an Icon...}
\usage{asIcon(x)
}
\description{Mark a character as an Icon}
\details{We use the class "Icon" to mark icons, which are css classnames or URLs\code{asIcon}: We use the class "Icon" to mark icons, which are css classnames or URLs


}
\value{same object with new class\code{asIcon}: same object with new class


}
\arguments{\item{x}{a character vector to mark with additional class of "Icon"}
}

\alias{asIcon}

