% Generated by roxygen2 (4.0.2): do not edit by hand
\name{asURL}
\alias{asURL}
\title{Add URL to class of object if not already}
\usage{
asURL(x)
}
\arguments{
\item{x}{object to add class to. Should be length 1 character}
}
\value{
returns object
}
\description{
Add URL to class of object if not already
}

