% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{blockHandler}
\alias{blockHandler}
\title{temporariliy block handler by id}
\usage{
blockHandler(x, cbid)
}
\arguments{
\item{x}{the object handler is called on.}

\item{cbid}{the callback id returned by an addHandlerXXX call}
}
\description{
temporariliy block handler by id
}
\note{
see the reference method \code{block_handlers} to block all
handlers for the widget.
}

