\name{blockHandler}
\alias{blockHandler}
\title{temporariliy block handler by id...}
\usage{blockHandler(x, cbid)
}
\description{temporariliy block handler by id}
\value{NULL\code{blockHandler}: NULL


}
\arguments{\item{x}{the object handler is called on.}
\item{cbid}{the callback id returned by an addHandlerXXX call}
}

\alias{blockHandler}

