% Generated by roxygen2 (4.0.2): do not edit by hand
\name{[<-.GComponent}
\alias{[<-.GComponent}
\title{assignment method for "["}
\usage{
\method{[}{GComponent}(x, i, j, ...) <- value
}
\arguments{
\item{x}{objecct}

\item{i}{row}

\item{j}{column}

\item{...}{passed to \code{set_items}}

\item{value}{passed to \code{set_items}}
}
\description{
assignment method for "["
}

