% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\name{column_editor}
\alias{column_editor}
\title{Generic function to create column editor for a table widget}
\usage{
column_editor(x, ...)
}
\arguments{
\item{x}{type of object}

\item{...}{ignored}
}
\description{
Generic function to create column editor for a table widget
}

