% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidget-proxy.R
\name{column_renderer}
\alias{column_renderer}
\title{Generic function to render a column in ExtJS Grid}
\usage{
column_renderer(x)
}
\arguments{
\item{x}{object to get column class from}
}
\description{
Generic function to render a column in ExtJS Grid
}

