% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{dim.GComponent}
\alias{dim.GComponent}
\title{dim method for GComponent's}
\usage{
\method{dim}{GComponent}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed along to \code{len} method}
}
\value{
dim of object, loosely interpreted
}
\description{
Return size of component, when rectangular size makes sense
}

