% Generated by roxygen2 (4.0.2): do not edit by hand
\name{editable<-}
\alias{editable<-}
\alias{editable<-.GComponent}
\title{Toggle editability of the object, if supported}
\usage{
editable(x, ...) <- value

\method{editable}{GComponent}(x, ...) <- value
}
\arguments{
\item{x}{object. Calls objects \code{set_editable} method}

\item{...}{passed to \code{set_editable} method}

\item{value}{logical}
}
\description{
Toggle editability of the object, if supported

assignment method for editable
}

