% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{enabled<-}
\alias{enabled<-}
\alias{enabled<-.GComponent}
\title{Set if widget is enabled}
\usage{
enabled(x) <- value

\method{enabled}{GComponent}(x) <- value
}
\arguments{
\item{x}{widget}

\item{value}{logical}
}
\description{
Set if widget is enabled

assignment method for enabled
}

