\name{escapeSingleQuote}
\alias{escapeSingleQuote}
\title{escape single quote...}
\usage{escapeSingleQuote(x)
}
\description{escape single quote}
\details{Useful as we often single quote arguments to the Ext functions}
\value{string with "'" replaced by "\'"}
\arguments{\item{x}{string to escape}
}

