% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ext-misc.R
\docType{class}
\name{ExtArgs-class}
\alias{ExtArgs}
\alias{ExtArgs-class}
\title{a class to handle map between R and arguments for constructor}
\description{
The ExtArgs are simply an Array instance wrapper with an extend method and
a conversion method
}
\section{Methods}{

\describe{
\item{\code{extend(l, overwrite = TRUE)}}{Extend argument list by list l.}

\item{\code{to_js_object()}}{Convert to string of object literal}
}}

