% Generated by roxygen2 (4.0.2): do not edit by hand
\name{focus<-}
\alias{focus<-}
\alias{focus<-.default}
\title{Set focus onto object.}
\usage{
focus(x) <- value

\method{focus}{default}(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{logical. Set focus state.}
}
\description{
For some widgets, this sets user focus (e.g. gedit gets focus for
typing).

Basic S3 method for focus
}

