\name{font<-}
\alias{font<-}
\alias{font<-.GComponent}
\title{Set font properties, if implemented}
\arguments{
  \item{x}{widget}

  \item{value}{font specification, a named list, eg.
  list(size=10, color="red")}
}
\description{
  Set font properties, if implemented

  assignment method for font
}

