% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{font<-}
\alias{font<-}
\alias{font<-.GComponent}
\title{Set font properties, if implemented}
\usage{
font(x) <- value

\method{font}{GComponent}(x) <- value
}
\arguments{
\item{x}{widget}

\item{value}{font specification, a named list, eg. list(size=10, color="red")}
}
\description{
Set font properties, if implemented

assignment method for font
}

