% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gaction}
\alias{gaction}
\title{gaction implementation}
\usage{
gaction(label, tooltip = label, icon = NULL, key.accel = NULL, handler,
  parent, ...)
}
\arguments{
\item{label}{Text for action}

\item{tooltip}{tooltip. Ignored for this toolkit}

\item{icon}{action icon class}

\item{key.accel}{keyboard accelerator. Single key, e.g. "X",
"LEFT" (arrow), "PAGE_UP", "Ctrl-n", "Alt-X". Use "Shift" to force
that. List of key events is given here:
\code{http://docs.sencha.com/ext-js/4-1/#!/api/Ext.EventObject}.}

\item{handler}{function called when action is invoked}

\item{parent}{toplevel window of action (where it can be
called). Required here}

\item{...}{ignored}
}
\value{
a \code{GWidget} object
}
\description{
actions are reusable encapsulations of actions. The
\code{svalue<-} method can be used to update the text associated
with an action. Use \code{enabled<-} to adjust sensitivity to
mouse events. The \code{visible<-} method can adjust if objects
proxying the action are visible. The \code{set_icon} reference
class method can be used to set the icon (no S3 method)
}
\details{
See the method \code{add_keybinding} to add a simple keybinding to
initiate this action.
}
\examples{
w <- gwindow()
a <- gaction("some action", handler=function(h,...) galert("asdf", parent=w), parent=w)
b <- gbutton(action=a, cont=w)
enabled(a) <- FALSE
svalue(a) <- "new text"
#
}

