% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gdialogs.R
\name{galert}
\alias{galert}
\title{quick alert message -- not modal or obtrusive (dropped from above in extjs)}
\usage{
galert(message, title = "message", delay = 3, parent)
}
\arguments{
\item{message}{message to display}

\item{title}{title of message}

\item{delay}{delay in seconds}

\item{parent}{parent window, typically gwindow instance. Necessary.}
}
\value{
not used
}
\description{
quick alert message -- not modal or obtrusive (dropped from above in extjs)
}
\examples{
w <- gwindow()
b <- gbutton("click me", cont=w, handler=function(h,...) {
galert("That hurt", parent=w)
})
}

