% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gbutton}
\alias{gbutton}
\title{Basic button widget}
\usage{
gbutton(text = "", handler = NULL, action = NULL, container = NULL, ...,
  width = NULL, height = NULL, ext.args = NULL)
}
\arguments{
\item{text}{button text}

\item{action}{passed to callback as \code{h$action}}

\item{handler}{optional means (to \code{\link{addHandlerChanged}})
to specify a handler for the default signal. A handler is a
function with signature \code{(h,...)} where \code{h} is a list
with components \code{obj} referring to the object emitting the
signal, \code{action} containing values passed to the
\code{action} argument, and possible other values.}

\item{container}{A parent container. In \pkg{gWidgetsWWW2} a
parent container is not optional (though it can be substituted
with the \code{parent} argument in some circumstances). The parent
specifies the widget heirarchy and the \code{...} argument is used
to pass along arguments to layout the child component in the
parent container. Typically, these are passed to the \code{add}
method of the parent container.}

\item{...}{Used to pass along argument to the parent container's
\code{add} method and possible other arguments to the underlying
ference class constructors.}

\item{width}{width in pixels of component. Sizing in
\pkg{gWidgetsWWW2} is sometimes necessary as the arguments
\code{expand} and \code{fill} are not well implemented.}

\item{height}{height in pixels of the component.}

\item{ext.args}{The contructors of \pkg{gWidgetsWWW2} ultimately
call an Ext constructor. The options passed to the Ext constructor
may be added to or overridden by use of this argument. Values are
passed in as named list components and with values converted into JavaScript
objects by \code{asJSObject}.}
}
\value{
a \code{GButton} instance
}
\description{
A button responds to mouse clicks by calling its handler
}
\examples{
w  <- gwindow("test")
sb <- gstatusbar("Powered by gWidgetsWWW and Rook", cont=w)
g <- ggroup(cont=w, horizontal=FALSE)
b <- gbutton("click me", cont=g, handler=function(h,...) {
  galert("hello world", parent=w)
})
## has an icon
b <- gbutton("up", cont=g)
## just an icon using ext override to remove text but leave icon
b <- gbutton("help", cont=g); b$call_Ext("setText")
## an action
a <- gaction("help", parent=w, handler=function(h,...) {
 galert("action", parent=w)
})
b <- gbutton(action=a, cont=g)
}

