% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gcodemirror}
\alias{gcodemirror}
\title{codemirror}
\usage{
gcodemirror(text = "", container = NULL, ..., width = NULL,
  height = NULL, ext.args = list())
}
\arguments{
\item{text}{initial text.}

\item{container}{A parent container. In \pkg{gWidgetsWWW2} a
parent container is not optional (though it can be substituted
with the \code{parent} argument in some circumstances). The parent
specifies the widget heirarchy and the \code{...} argument is used
to pass along arguments to layout the child component in the
parent container. Typically, these are passed to the \code{add}
method of the parent container.}

\item{...}{Used to pass along argument to the parent container's
\code{add} method and possible other arguments to the underlying
ference class constructors.}

\item{width}{width in pixels of component. Sizing in
\pkg{gWidgetsWWW2} is sometimes necessary as the arguments
\code{expand} and \code{fill} are not well implemented.}

\item{height}{height in pixels of the component.}

\item{ext.args}{The contructors of \pkg{gWidgetsWWW2} ultimately
call an Ext constructor. The options passed to the Ext constructor
may be added to or overridden by use of this argument. Values are
passed in as named list components and with values converted into JavaScript
objects by \code{asJSObject}.}
}
\value{
a GCodeMirror reference class
}
\description{
A gtext object modified to use codemirror for syntax highlighting
}
\details{
The CodeMirror project (codemirror.net) provides a code editor in
javascript with an R mode. This widget incorporates that into
\pkg{gWidgetsWWW2}. This does R syntax highlighting. To use
codemirror with gtext (and no highlighting), pass the argument
\code{use.codemirror=TRUE} to the \code{gtext} call.
}
\author{
john verzani
}

