\name{gcombobox}
\alias{gcombobox}
\title{Alternative name for gcomobox...}
\description{Alternative name for gcomobox}
\details{\code{gcombobox}: The name gdroplist is an alternative (not encouraged) for gcomobbox

\code{gcombobox}: The name gdroplist is an alternative (not encouraged) for gcomobbox


}
\alias{gcombobox}

