% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gcontainer.R
\docType{data}
\name{GContainer}
\alias{GContainer}
\title{Base reference class for container objects.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x252f290> 
 - attr(*, "name")= chr "GContainer_generator"
}}
\usage{
GContainer
}
\description{
Container objects are those that hold other components and may be
passed to constructors through their \code{container}
argument. All objects keep a reference to their parent container,
containers also keep references to their child components. See the
\code{children} array, which is used for this.
}
\details{
The main method for containers is the \code{add} method, though
this is rarely called, as the underlying method is used in a
widget's constructor when a parent container is specified through
the standard \code{container} argument.

When an object is added one may need to recompute the layout. The
\code{do_layout} method will inittiate this. To have this done
each time a child is added set the argument \code{auto_layout} to
\code{TRUE}.
}
\keyword{datasets}

