% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gdf}
\alias{gdf}
\title{A widget for editing a data frame}
\usage{
gdf(items = NULL, handler = NULL, action = NULL, container = NULL, ...,
  width = 300, height = 200, ext.args = NULL,
  name = deparse(substitute(items)))
}
\arguments{
\item{items}{data frame to be edited}

\item{name}{name of data frame appearing in title}

\item{handler}{optional means (to \code{\link{addHandlerChanged}})
to specify a handler for the default signal. A handler is a
function with signature \code{(h,...)} where \code{h} is a list
with components \code{obj} referring to the object emitting the
signal, \code{action} containing values passed to the
\code{action} argument, and possible other values.}

\item{action}{Passed to handler to parameterize a call}

\item{container}{A parent container. In \pkg{gWidgetsWWW2} a
parent container is not optional (though it can be substituted
with the \code{parent} argument in some circumstances). The parent
specifies the widget heirarchy and the \code{...} argument is used
to pass along arguments to layout the child component in the
parent container. Typically, these are passed to the \code{add}
method of the parent container.}

\item{...}{Used to pass along argument to the parent container's
\code{add} method and possible other arguments to the underlying
ference class constructors.}

\item{width}{width in pixels of component. Sizing in
\pkg{gWidgetsWWW2} is sometimes necessary as the arguments
\code{expand} and \code{fill} are not well implemented.}

\item{height}{height in pixels of the component.}

\item{ext.args}{The contructors of \pkg{gWidgetsWWW2} ultimately
call an Ext constructor. The options passed to the Ext constructor
may be added to or overridden by use of this argument. Values are
passed in as named list components and with values converted into JavaScript
objects by \code{asJSObject}.}
}
\value{
a \code{GDf} reference class object.
}
\description{
A widget for editing a data frame This implementation allows the
user to edit a row (row-by-row). Change handlers are called
after an edit is performed. The user can add a new row or remove
the currently selected row (if more than 1 is present). The hidden
argument \code{do_add_remove_buttons} can be set to \code{FALSE}
to prevent this option. For large data sets, paging is used. the
hidden argument \code{page_size} (with default of 200) can be set
to adjust the size of each page.
}
\author{
john verzani
}

