% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_tempfile_url}
\alias{get_tempfile_url}
\title{Get url for a tempfile created by \code{get_tempfile}}
\usage{
get_tempfile_url(f)
}
\arguments{
\item{f}{a file name produced by \code{get_tempfile}}
}
\description{
Requires the tempdir to be mapped to a specific url
}
\examples{
f <- get_tempfile(ext=".svg")
get_tempfile_url(f)
}

