% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getStockIconByName}
\alias{getStockIconByName}
\title{Get the icon class (or url) by its name}
\usage{
getStockIconByName(icon_name, css = TRUE)
}
\arguments{
\item{icon_name}{icon name to be looked up in stock icons}

\item{css}{logical. If TRUE, return css class name, else url}
}
\value{
Icon css class (or url) or NA if not there
}
\description{
Get the icon class (or url) by its name
}
\examples{
getStockIconByName("up") ## "x-gw-icon-up"
getStockIconByName("Not there") ## NULL
}

