\name{getStockIcons}
\alias{getStockIcons}
\title{Return stock icons...}
\description{Return stock icons}
\details{These appear as a list with a name, CSS class (setIconClass), and url (setIcon)\code{getStockIcons}: These appear as a list with a name, CSS class (setIconClass), and url (setIcon)


}
\value{a list of icons. Unlike other gWidgets
implementations. Each component is alist with css name and url, as
both are useful.\code{getStockIcons}: a list of icons. Unlike other gWidgets
implementations. Each component is alist with css name and url, as
both are useful.


}

\alias{getStockIcons}

