% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ghtml.R
\name{ghtml}
\alias{ghtml}
\title{widget to render HTML text pages}
\usage{
ghtml(x, container = NULL, ..., width = NULL, height = NULL,
  ext.args = NULL)
}
\arguments{
\item{x}{an HTML string or a URL.}

\item{container}{parent container}

\item{...}{passed to add method of parent container}

\item{width}{width of widge in pixels}

\item{height}{height of widget in pixels}

\item{ext.args}{extra arguments to pass to constructor}
}
\value{
an \code{ExtWidget} object
}
\description{
widget to render HTML text pages
}
\examples{
w <- gwindow()
g <- ggroup(cont=w, horizontal=FALSE)
h <- ghtml("<b>this is bold</b>", cont=g)
}

