% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gmenu}
\alias{gmenu}
\title{Menubar implementation}
\usage{
gmenu(menulist, popup = FALSE, action = NULL, container = NULL, ...,
  ext.args = NULL)
}
\arguments{
\item{menulist}{list of actions. Actions must have parent specified}

\item{popup}{Logical. ignored for now}

\item{action}{Passed to handler to parameterize a call}

\item{container}{A parent container. In \pkg{gWidgetsWWW2} a
parent container is not optional (though it can be substituted
with the \code{parent} argument in some circumstances). The parent
specifies the widget heirarchy and the \code{...} argument is used
to pass along arguments to layout the child component in the
parent container. Typically, these are passed to the \code{add}
method of the parent container.}

\item{...}{Used to pass along argument to the parent container's
\code{add} method and possible other arguments to the underlying
ference class constructors.}

\item{ext.args}{The contructors of \pkg{gWidgetsWWW2} ultimately
call an Ext constructor. The options passed to the Ext constructor
may be added to or overridden by use of this argument. Values are
passed in as named list components and with values converted into JavaScript
objects by \code{asJSObject}.}
}
\value{
an ExtWidget object
}
\description{
A menubar for gwindow instances
%
Menu items are specified with a named list. The heirarchical
nature of the list maps to the heirarchical menu structure, with
the names giving the menu names. The menu actions are specified
using \code{gaction} elements. These may also be
\code{gseperator()} instances (no parent necessary here).
%
Menubars should only be added to \code{gwindow} instances, but
this is not strictly enforced, any \code{Ext.Panel}-based
container would do.
}
\examples{
w <- gwindow()
sb <- gstatusbar("Powered by gWidgetsWWW and Rook", cont=w)
stub = function(...) galert("stub", parent=w)
l <- list(
File = list(open=gaction("Open", handler=stub, parent=w),
            new = gaction("New", handler=stub, parent=w),
            gseparator(),
            quit = gaction("Quit", handler=stub, parent=w)),
Edit = list(save = gaction("Save", handler=stub, parent=w))
)
m <- gmenu(l, cont=w)
gtext("Text goes here", cont=w)
}

