% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdialogs.R
\name{gmessage}
\alias{gmessage}
\title{A simple message dialog.}
\usage{
gmessage(message, title = "message", icon = NULL, parent = NULL, ...)
}
\arguments{
\item{message}{main message.}

\item{title}{Title for dialog's window}

\item{icon}{icon to decorate dialog. One of \code{c("info", "warning", "error", "question")}.}

\item{parent}{parent container (the main window instance)}

\item{...}{ignored}
}
\value{
return value ignored
}
\description{
A simple message dialog.
}
\examples{
w <- gwindow()
gbutton("click me for a message", cont=w, handler=function(h,...) {
gmessage("Hello there", parent=w)
})
}

