% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gpanedgroup}
\alias{gpanedgroup}
\title{A paned group container}
\usage{
gpanedgroup(horizontal = TRUE, container = NULL, ..., width = NULL,
  height = 400, ext.args = NULL, collapsible = FALSE,
  default.size = NULL)
}
\arguments{
\item{horizontal}{logical. Side by side (or top/bottom) layout}

\item{container}{A parent container. In \pkg{gWidgetsWWW2} a
parent container is not optional (though it can be substituted
with the \code{parent} argument in some circumstances). The parent
specifies the widget heirarchy and the \code{...} argument is used
to pass along arguments to layout the child component in the
parent container. Typically, these are passed to the \code{add}
method of the parent container.}

\item{...}{Used to pass along argument to the parent container's
\code{add} method and possible other arguments to the underlying
ference class constructors.}

\item{width}{width in pixels of component. Sizing in
\pkg{gWidgetsWWW2} is sometimes necessary as the arguments
\code{expand} and \code{fill} are not well implemented.}

\item{height}{height in pixels of the component.}

\item{ext.args}{The contructors of \pkg{gWidgetsWWW2} ultimately
call an Ext constructor. The options passed to the Ext constructor
may be added to or overridden by use of this argument. Values are
passed in as named list components and with values converted into JavaScript
objects by \code{asJSObject}.}

\item{collapsible}{logical. If TRUE, one child will have a button so that it can be collapsed}

\item{default.size}{Integer pixel size of right or bottom
container when drawn. Defaults to half of height or width, or 200
if those are NULL. There is no \code{svalue<-} method to adjust the size}
}
\value{
an ExtContainer object
}
\description{
This container has two children and a vertical/horizontal sash to
allocate space between them. Simply use as a parent container
twice. Unlike other gWidgets implementations, \code{svalue<-} is not implemented.
If in horizontal mode, the second widget added will be to the left.
}
\examples{
w <- gwindow()
pg <- gpanedgroup(cont=w, horizontal=TRUE)
gbutton("left", cont=pg)
gbutton("right", cont=pg)
}

