% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gstatusbar.R
\name{gstatusbar}
\alias{gstatusbar}
\title{Status bar for gwindow instances}
\usage{
gstatusbar(text = "", container = NULL, ..., ext.args = NULL)
}
\arguments{
\item{text}{text for label}

\item{container}{A parent container. In \pkg{gWidgetsWWW2} a
parent container is not optional (though it can be substituted
with the \code{parent} argument in some circumstances). The parent
specifies the widget heirarchy and the \code{...} argument is used
to pass along arguments to layout the child component in the
parent container. Typically, these are passed to the \code{add}
method of the parent container.}

\item{...}{Used to pass along argument to the parent container's
\code{add} method and possible other arguments to the underlying
ference class constructors.}

\item{ext.args}{The contructors of \pkg{gWidgetsWWW2} ultimately
call an Ext constructor. The options passed to the Ext constructor
may be added to or overridden by use of this argument. Values are
passed in as named list components and with values converted into JavaScript
objects by \code{asJSObject}.}
}
\value{
an ExtWidget instance
}
\description{
Status for main window. Use \code{gwindow} instance for parent
container. The \code{svalue<-} method can be used to change the
value.
}
\examples{
w <- gwindow()
sb <- gstatusbar("Powered by gWidgetsWWW and Rook")
}

