% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gsvg.R
\docType{class}
\name{gsvg-class}
\alias{GSvg}
\alias{gsvg-class}
\title{base class for gsvg}
\description{
base class for gsvg
}
\section{Methods}{

\describe{
\item{\code{get_svg_id()}}{Return id of svg div element}

\item{\code{get_tempfile(ext = ".txt")}}{Create a new temporary file}

\item{\code{set_value(value, ...)}}{Set main property, invoke change handler on change}
}}

