% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GWidgetHTMLProxy-class}
\alias{GWidgetHTMLProxy}
\alias{GWidgetHTMLProxy-class}
\title{Proxy for ghtml}
\description{
Proxy for ghtml
}
\section{Methods}{

\describe{
\item{\code{get_json_data(...)}}{The params determine what to pass back. }
}}

