% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-toplevel.R
\docType{class}
\name{GWidgetsTopLevel-class}
\alias{GWidgetsTopLevel}
\alias{GWidgetsTopLevel-class}
\title{A toplevel object for a gWidgetsWWW2 app to store page-specific values}
\format{An \code{\link{R6Class}} generator object}
\usage{
GWidgetsTopLevel
}
\description{
The toplevel object for gWidgetsWWW2 apps stores page specific
information. There can potentially be several \code{gwindow}
instances on a page (using \code{renderTo}), however all share the
same toplevel instance. This instance holds a reference to the
evaluation environment, the request that generates the page, the
objects populating the page, and a queue for holding JavaScript
commands (produced by method calls, say).
}
\details{
The sharing of this toplevel object among each component is the
reason why the constructors require a specification of either a
\code{container} or \code{parent} argument.
}
\section{Fields}{

\describe{
\item{\code{e}}{The evaluation environment}

\item{\code{request}}{The request}

\item{\code{objects}}{Objects on the page as an \code{\link{Array}} object}

\item{\code{js_queue}}{JavaScript queue to flush as an \code{\link{Array}} object}

\item{\code{js_queue}}{JavaScript queue to flush as an \code{\link{Array}} object}

\item{\code{cookies}}{An environment holding cookies}

\item{\code{do_layout_cmd}}{The command used for the layout}
}}
\keyword{datasets}

