% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{insert.GComponent}
\alias{insert.GComponent}
\title{method for insert}
\usage{
\method{insert}{GComponent}(x, value, where = "end", font.attr = NULL,
  do.newline = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{value}{value to insert}

\item{where}{to insert}

\item{font.attr}{ignored}

\item{do.newline}{logical}

\item{...}{ignored}
}
\description{
method for insert
}

