% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{insert}
\alias{insert}
\title{Insert text into \code{gtext}}
\usage{
insert(x, value, where = "end", font.attr = NULL, do.newline = TRUE, ...)
}
\arguments{
\item{x}{object}

\item{value}{text value}

\item{where}{where to insert \code{c("end", "beginning", "at.cursor")}}

\item{font.attr}{ignored. Font attribute for new text}

\item{do.newline}{Do we add a new line?}

\item{...}{ignored}
}
\value{
no useful return value
}
\description{
Used to insert text into \code{gtext} insances
}

