\name{insert}
\alias{insert}
\title{Insert text into gtext...}
\usage{insert(obj, value, where="end", font.attr, do.newline=TRUE, ...)
}
\description{Insert text into \code{gtext}}
\details{Used to insert text into \code{gtext} insances}
\value{no useful return value}
\arguments{\item{obj}{object}
\item{value}{text value}
\item{where}{where to insert \code{c("end", "beginning", "at.cursor")}}
\item{font.attr}{ignored. Font attribute for new text}
\item{do.newline}{Do we add a new line?}
\item{...}{ignored}
}

