\name{isURL}
\alias{isURL}
\title{Is value a URL: either of our class URL or matches url string: ftp://, http:// or file:///...}
\usage{isURL(x)
}
\description{Is value a URL: either of our class URL or matches url string: ftp://, http:// or file:///}
\value{Logical indicating if a URL.\code{isURL}: Logical indicating if a URL.


}
\arguments{\item{x}{length 1 character value to test}
}

\alias{isURL}

