% Generated by roxygen2 (4.0.1): do not edit by hand
\name{names<-.GComponent}
\alias{names<-.GComponent}
\alias{namesGComponent}
\title{assignment method for names}
\usage{
\method{names}{GComponent}(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{new names}
}
\description{
assignment method for names
}

