% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeHandler}
\alias{removeHandler}
\title{Remove handler}
\usage{
removeHandler(x, cbid)
}
\arguments{
\item{x}{the object handler is called on.}

\item{cbid}{the callback id returned by an addHandlerXXX call}
}
\description{
Remove handler
}
\note{
see the reference method \code{remove_handlers} to remove
all handlers on a widget.
}

