\name{removeHandler}
\alias{removeHandler}
\title{Remove handler...}
\usage{removeHandler(x, cbid)
}
\description{Remove handler}
\value{NULL\code{removeHandler}: NULL


}
\arguments{\item{x}{the object handler is called on.}
\item{cbid}{the callback id returned by an addHandlerXXX call}
}

\alias{removeHandler}

