% Generated by roxygen2 (4.0.1): do not edit by hand
\name{size<-}
\alias{size<-}
\alias{size<-.GComponent}
\title{Set size property, if implemented}
\usage{
size(x, ...) <- value

\method{size}{GComponent}(x, ...) <- value
}
\arguments{
\item{x}{widget}

\item{...}{passed on}

\item{value}{size specification, for most widgets a pair c(width, height), but can have exceptions}
}
\description{
Set size property, if implemented

assignment method for size
}

