% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/utils.R
\name{String_render}
\alias{String_render}
\title{Internal function to call whisker.render and compose with String}
\usage{
String_render(tpl, l)
}
\arguments{
\item{tpl}{template}

\item{l}{data}
}
\value{
a character containing filled in template
}
\description{
Internal function to call whisker.render and compose with String
}

