% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{svalue}
\alias{svalue}
\alias{svalue.GComponent}
\title{Return main value associated with a widget}
\usage{
svalue(x, index = NULL, drop = NULL, ...)

\method{svalue}{GComponent}(x, index = NULL, drop = NULL, ...)
}
\arguments{
\item{x}{the widget}

\item{index}{if specified as \code{TRUE} calls \code{get_index}, else}

\item{drop}{passed along, in many cases used like \code{drop} call for \code{[}.
\code{get_value} reference methods.}

\item{...}{passed to \code{get_value} or \code{get_index}
method. May include arguments \code{index} or \code{drop}}
}
\description{
Return main value associated with a widget

svalue method
}

