\name{svalue<-}
\alias{svalue<-}
\alias{svalue<-.GComponent}
\title{Set main value associated with a widget}
\usage{
  svalue(x, ..., value) <- value

  \method{"svalue"}{GComponent} (x, index = NULL, ...,
    value) <- value
}
\arguments{
  \item{x}{object}

  \item{...}{passed to \code{set_value} method. May include
  arguments for \code{index}}

  \item{value}{value to set}
}
\description{
  Set main value associated with a widget

  assignment method for svalue
}

