\name{toJSArray}
\alias{toJSArray}
\title{Make a JS array from an R object...}
\usage{toJSArray(x, doBrackets=TRUE)
}
\description{Make a JS array from an R object}
\value{JSON encoded value from arrary\code{toJSArray}: JSON encoded value from arrary


}
\arguments{\item{x}{R object to make into an array}
\item{doBrackets}{logical Use brackets in ouput []}
}

\alias{toJSArray}

