\name{visible<-}
\alias{visible<-}
\alias{visible<-.GComponent}
\title{Primarily used to set if widget is shown, but also has other meanings}
\arguments{
  \item{x}{object Calls objects \code{set_visible} method}

  \item{value}{logical}
}
\description{
  Primarily used to set if widget is shown, but also has
  other meanings

  assignment method for visible
}

