% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/gwidgets-methods.R
\name{visible<-}
\alias{visible<-}
\alias{visible<-.GComponent}
\title{Primarily used to set if widget is shown, but also has other meanings}
\usage{
visible(x) <- value

\method{visible}{GComponent}(x) <- value
}
\arguments{
\item{x}{object Calls objects \code{set_visible} method}

\item{value}{logical}
}
\description{
Primarily used to set if widget is shown, but also has other meanings

assignment method for visible
}

